Remove Deploy from CLI (#300)
* Moving CLI elements
