+++ cmd/chainbridge/flags.go
@@ -54,61 +54,10 @@ var (
 	}
 )
 
-// Deploy Contracts Flags
-var (
-	PortFlag = cli.StringFlag{
-		Name:  "port",
-		Usage: "The port at which your local chain instance or ganache running",
-		Value: "8545",
-	}
-	NumRelayersFlag = cli.IntFlag{
-		Name:  "validators",
-		Usage: "Specify total number of validators",
-		Value: 2,
-	}
-	RelayerThresholdFlag = cli.IntFlag{
-		Name:  "validatorthreshold",
-		Usage: "Specify validator threshold",
-		Value: 2,
-	}
-	PKFlag = cli.StringFlag{
-		Name:  "pk",
-		Usage: "Specify private key of account you wish to deploy contracts from",
-		Value: "000000000000000000000000000000000000000000000000000000416c696365",
-	}
-	MinCountFlag = cli.IntFlag{
-		Name:  "mincount",
-		Usage: "Specify min count for bridge asset",
-		Value: 10,
-	}
-)
-
 // Test Setting Flags
 var (
 	TestKeyFlag = cli.StringFlag{
 		Name:  "testkey",
 		Usage: "Applies a predetermined test keystore to the chains",
-	}
-)
-
-//Ganache Flags
-var (
-	ResetFlag = cli.BoolFlag{
-		Name:  "reset",
-		Usage: "Removes the default keyring flags, used if a replacement is wanted",
-	}
-	MnemFlag = cli.StringFlag{
-		Name:  "mnemonic",
-		Usage: "Allows a custom mnemonic to be inputted",
-		Value: "",
-	}
-	AccountFlag = cli.StringFlag{
-		Name:  "account",
-		Usage: "Allows a user to add a test account to the Ganache Instance. Input as one comma seperated string",
-	}
-	AmountFlag = cli.StringFlag{
-		Name:  "amount",
-		Usage: "Sets the starting amount of eth for each account",
-		Value: "100000000000000000000",
 	}
 )
